requirements.txt
python-dotenv==1.0.1
pandas==2.2.3
openai==1.51.0
streamlit==1.39.0
plotly==5.24.1
matplotlib==3.9.2
seaborn==0.13.2
statsmodels==0.14.4
fastapi==0.115.2
uvicorn==0.15.0
jinja2==3.0.2
pydantic==2.9.2